Prevent divide by zero errors in temp calculation
If a black object or if it is low light causes R2 in _temperature_and_lux_dn40 to be zero the line for color temperature calculation will throw a divide by zero error. This can be repeated by running tcs34725_simpletest.py which prints lux and temp, disabling the LED by pulling LED pin to GND and placing something on the sensor to block ambient light (e.g. your finger). This can also occur, although less reliably, with a very black object with the LED on.

Proposed fix is just an if to check if R2 is greater then zero. If it isn't use a small value in it's place to allow B2 to still be considered.